# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workspaces": {
      "get": {
        "operationId": "WorkspacesController_listWorkspaces",
        "parameters": [
          {
            "name": "x-task-user-id",
            "in": "header",
            "description": "Temporary trusted user context header until AuthModule owns request identity. Not an authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceSummaryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List workspaces visible to the current user",
        "tags": ["workspaces"]
      }
    },
    "/workspaces/{workspaceId}": {
      "get": {
        "operationId": "WorkspacesController_getWorkspace",
        "parameters": [
          {
            "name": "x-task-user-id",
            "in": "header",
            "description": "Temporary trusted user context header until AuthModule owns request identity. Not an authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Workspace is missing or not visible to the current user."
          }
        },
        "summary": "Get one workspace with members",
        "tags": ["workspaces"]
      }
    },
    "/workspaces/{workspaceId}/members": {
      "get": {
        "operationId": "WorkspacesController_listMembers",
        "parameters": [
          {
            "name": "x-task-user-id",
            "in": "header",
            "description": "Temporary trusted user context header until AuthModule owns request identity. Not an authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceMemberDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Workspace is missing or not visible to the current user."
          }
        },
        "summary": "List members for one visible workspace",
        "tags": ["workspaces"]
      }
    },
    "/workspaces/{workspaceId}/members/{memberId}/role": {
      "patch": {
        "operationId": "WorkspacesController_updateMemberRole",
        "parameters": [
          {
            "name": "x-task-user-id",
            "in": "header",
            "description": "Temporary trusted user context header until AuthModule owns request identity. Not an authentication mechanism.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemberDto"
                }
              }
            }
          },
          "400": {
            "description": "Workspace member role payload is invalid."
          },
          "403": {
            "description": "Current user cannot update this workspace member role."
          },
          "404": {
            "description": "Workspace member is missing or not visible to the current user."
          }
        },
        "summary": "Update one workspace member role",
        "tags": ["workspaces"]
      }
    },
...
  },
  "info": {
    "title": "tAsk API",
    "description": "Backend API contract for tAsk.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "WorkspaceSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Studio"
          },
          "slug": {
            "type": "string",
            "example": "studio"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "name", "slug", "createdAt", "updatedAt"]
      },
      "WorkspaceMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": ["owner", "admin", "member", "guest"]
          },
          "displayName": {
            "type": "string",
            "example": "Alex"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "alex@example.com"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/avatar.png"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "workspaceId", "userId", "role", "displayName", "createdAt", "updatedAt"]
      },
      "WorkspaceDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Studio"
          },
          "slug": {
            "type": "string",
            "example": "studio"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceMemberDto"
            }
          }
        },
        "required": ["id", "name", "slug", "createdAt", "updatedAt", "members"]
      },
      "UpdateWorkspaceMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": ["admin", "member", "guest"]
          }
        },
        "required": ["role"]
      },
...
    }
  }
}